# docs/swagger.yaml
definitions:
  requests.LoginReq:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  requests.ProfileResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        $ref: '#/definitions/requests.RoleResponse'
      role_id:
        type: string
      verified_at:
        type: string
    type: object
  requests.RegistRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  requests.RoleResponse:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
    type: object
info:
  contact: {}
paths:
  /Auth/Login:
    post:
      consumes:
      - application/json
      description: login by username and password
      parameters:
      - description: Login data
        in: body
        name: LoginData
        required: true
        schema:
          $ref: '#/definitions/requests.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Login
      tags:
      - auth
  /Auth/Profile:
    get:
      consumes:
      - application/json
      description: get user profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/requests.ProfileResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Profile
      tags:
      - auth
  /Auth/Registration:
    post:
      consumes:
      - application/json
      description: register a new user
      parameters:
      - description: Register data
        in: body
        name: RegisterData
        required: true
        schema:
          $ref: '#/definitions/requests.RegistRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Register
      tags:
      - auth
  /Auth/Verification/{id}:
    get:
      consumes:
      - application/json
      description: verification by user id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Verification
      tags:
      - auth
  /FileUpload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file and return its path and URL
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: File Upload
      tags:
      - file
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"